requirements.txt
numpy
pandas
matplotlib

RPi.GPIO
pmw3901